move error handling into configs
